move selector to correct file
