+++ ui/shared/components/panel/variants/selectors.js
@@ -55,23 +55,6 @@ export const getIndividualGeneDataByFamilyGene = createSelector(
   ),
 )
 
-export const getIndividualPhenotypeGeneScores = createSelector(
-  getGenesById,
-  getPhenotypeGeneScoresByIndividual,
-  (genesById, phenotypeGeneScoresByIndividual) => (
-    Object.entries(phenotypeGeneScoresByIndividual || {}).reduce((acc, [individualGuid, dataByGene]) => ({
-      ...acc,
-      [individualGuid]: Object.entries(dataByGene).reduce((acc2, [geneId, dataByTool]) => ([
-        ...acc2,
-        ...Object.entries(dataByTool).reduce((acc3, [tool, data]) => ([
-          ...acc3,
-          ...data.map(d => ({ ...d, tool, gene: genesById[geneId], rowId: `${geneId}-${tool}-${d.diseaseId}` })),
-        ]), []),
-      ]), []),
-    }), {})
-  ),
-)
-
 // Saved variant selectors
 export const getSavedVariantTableState = state => (
   state.currentProjectGuid ? state.savedVariantTableState : state.allProjectSavedVariantTableState
